Filter by best possible RPM to copy (w/constraints)
Pulp's API will return the href of a collection of RPM names, but it
will return ALL the known versions of each named RPM in the repository.

This is no good when we are looking for the hrefs of specific RPMs to
copy into each "slim" repository, so this patch does two big things:

* Updates the RPM/repo YAML data structure to provide `name:` and
  an optional `version:` constraint (using comparisons from Gem::Version
  on the RPM versions)
* Adds logic to compare RPM versions and pick the best one (a little
  more nuanced; see the relevant comments)

This resulted in a collection of el7 6.5.0 repos with RPMs that are MUCH
closer to the original.