(SIMP-10622) Mirror repos with sync_policy
Before this patch, Pulp would fail to mirror EPEL8 with the message:

    This repository uses features which are incompatible with 'mirror'
    sync. Please sync without mirroring enabled.

This is due to a new version of the [pulp_rpm][0] plugin, which
deprecates `mirror=True`, which now refuses to mirror certain kinds repo
metadata.

This patch fixes the EPEL8 mirror by replacing the deprecated `mirror`
with the new `sync_policy`

It also adds minor sanitization tweaks for inferred dist base paths.

[SIMP-10622] #close